+++ src/pages/HomePage/index.tsx
@@ -1,10 +1,8 @@
-import { InfiniteData, QueryErrorResetBoundary, useQueryClient } from '@tanstack/react-query'
+import { QueryErrorResetBoundary } from '@tanstack/react-query'
 import { Suspense, useEffect } from 'react'
 import { ErrorBoundary } from 'react-error-boundary'
 import { Helmet } from 'react-helmet-async'
 import { useNavigate, useSearchParams } from 'react-router-dom'
-import { useWebSocket } from '~/WebSocketContext'
-import { FetchChatMessageListResponse } from '~apis/chat/fetchChatMessageList'
 import { useHomePageData } from '~apis/main/useHomePageData'
 import DogHand from '~assets/dog_hand.svg?react'
 import BellIcon from '~assets/icons/bell_icon.svg?react'
@@ -17,110 +15,16 @@ import Profile from '~components/Profile'
 import { Separator } from '~components/Separator'
 import { Typo14, Typo17, Typo24 } from '~components/Typo'
 import { FAMILY_ROLE } from '~constants/familyRole'
-import { queryKey } from '~constants/queryKey'
+import useSubscribe from '~hooks/useSubscribe'
 import NotificationModal from '~modals/NotificationModal'
 import { useModalStore } from '~stores/modalStore'
-import { APIResponse, CommonAPIResponse } from '~types/api'
 import * as S from './styles'
 
 function HomeContent() {
-  const { isConnected, subscribe } = useWebSocket()
   const { data } = useHomePageData()
   const { pushModal } = useModalStore()
-  const queryClient = useQueryClient()
 
-  useEffect(() => {
-    if (isConnected) {
-      console.log('구독!')
-      subscribe(`/user/queue/errors`, message => {
-        const response = JSON.parse(message.body)
-        console.log('에러 구독', response)
-      })
-
-      subscribe(`/sub/message/${data.email}`, message => {
-        const response = JSON.parse(message.body) as {
-          data: {
-            chatRoomId: number
-            unreadCount: number
-          }[]
-        }
-        console.log('이메일 구독', response)
-        response.data.forEach(chatRoom => {
-          subscribe(`/sub/chat/${chatRoom.chatRoomId}`, message => {
-            const res = JSON.parse(message.body) as APIResponse<
-              Pick<
-                CommonAPIResponse,
-                'chatId' | 'createdAt' | 'updatedAt' | 'chatRoomId' | 'memberInfo' | 'isRead' | 'text'
-              >
-            >
-            console.log('채팅방 구독', res)
-            queryClient.invalidateQueries({
-              queryKey: queryKey.social.chatRoomList(),
-            })
-            if (res.data.chatId)
-              queryClient.setQueryData<InfiniteData<APIResponse<FetchChatMessageListResponse>>>(
-                queryKey.social.chatMessageList(res.data.chatRoomId),
-                oldData => {
-                  if (!oldData) {
-                    const initialPage: APIResponse<FetchChatMessageListResponse> = {
-                      code: 200,
-                      status: 'OK',
-                      message: 'Success',
-                      data: {
-                        content: [res.data],
-                        size: 1,
-                        number: 0,
-                        numberOfElements: 1,
-                        first: true,
-                        last: true,
-                        empty: false,
-                        sort: {
-                          empty: true,
-                          sorted: false,
-                          unsorted: true,
-                        },
-                        pageable: {
-                          offset: 0,
-                          sort: {
-                            empty: true,
-                            sorted: false,
-                            unsorted: true,
-                          },
-                          pageSize: 1,
-                          paged: true,
-                          pageNumber: 0,
-                          unpaged: false,
-                        },
-                      },
-                    }
-                    return {
-                      pages: [initialPage],
-                      pageParams: [null],
-                    }
-                  }
-                  return {
-                    ...oldData,
-                    pages: oldData.pages.map((page, index) => {
-                      if (index === 0) {
-                        return {
-                          ...page,
-                          data: {
-                            ...page.data,
-                            content: [...page.data.content, res.data],
-                            numberOfElements: page.data.numberOfElements + 1,
-                          },
-                        }
-                      }
-                      return page
-                    }),
-                  }
-                }
-              )
-          })
-        })
-      })
-    }
-  }, [isConnected])
+  useSubscribe()
 
   return (
     <>
